[CWS] refacto
